Fixed issue #2
If a bishop is on the first or last rank it was able to move diagonally
because of the CoordinateIterator only being able to be created from a
start position to an end position. In the case of diagonally moving
pieces I used the corners to mark the end. Since the iterator calculates
deltas and uses those to iterate over the coordinate it thinks this is a
straight moving piece like a rook and allows these moves