Refactored all of move generation
This fix allowed us to remove the is_guarded function which created a
new board to test if the king is in check and stuff

This also fixes #6 which was caused by king move generations attacking
squares did say that one of the king could attack the shared squares
which caused the other to be able to walk there. Now we calculate
attacking and possible moves seperately so this is no longer an issue?