Fixes #11
Issue #11 was caused by the fix for issue #8. Issue #8 was fixed by
making the check for attacked and occupied board squares inclusive to
check the king square as well and since the attacked and occupied checks
are done in the same closure this did not take into account that the
king obviously always occupies it's slot when it is meant to castle. The
fix was simply ignoring the occupied check for the king file